Extract notification logic into NotificationStore
Moves checkForActionPrompt, playNotificationSound, and sendActionNotification
out of +page.svelte into src/lib/stores/notifications.svelte.ts, along with
their associated module-level state (outputBuffer, lastNotifyTime, soundRepeatInterval,
bufferTimer). Call site updated to use notifications.checkForActionPrompt(...).
